update to allow opening single item in new page
