+++ lore-search-app/app/item/[id]/page.tsx
@@ -1,0 +1,10 @@
+import { fetchLoreEntry } from "@/app/lib/data";
+import ItemPage from "../../ui/item/itemPage";
+
+export default async function Page({ params }: { params: { id: string } }) {
+  let data = await fetchLoreEntry(params.id);
+
+  return (
+    <div className="p-2">{data ? <ItemPage lore={data} /> : "Not Found"}</div>
+  );
+}
