--- lore-search-app/app/search/layout.tsx
@@ -1,10 +1,0 @@
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <div>
-      <main className="min-h-screen min-w-full">
-        {children}
-      </main>
-    </div>
-  );
-}
